Separated utility classes and functions in mini dancers in separated header files.
